pyproject.toml
[tool.poetry]
name = "ordinal-predictions"
version = "0.1.0"
description = ""
authors = ["Ethan Thoma <ethanthoma@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-bigquery = "^3.13.0"
pandas = "^2.1.3"
mord = "^0.7"
scipy = "^1.11.4"
scikit-learn = "^1.3.2"
google-cloud-storage = "^2.13.0"
pyarrow = "^14.0.1"
seaborn = "^0.13.0"
matplotlib = "^3.8.2"
statsmodels = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

